Port remainder of tools to Windows
Thanks to Erik G <erik@6809.org> for the patch.

Quoth Erik:

"As a thank you for putting up with my nonsense, I ported the remaining
binaries. Although I'm not using them (yet), I did verify that they work and
the changes are basically zero so I have confidence in them."

+++ lwlink/lwlink.h
@@ -80,7 +80,7 @@ struct fileinfo_s
 {
 	char *filename;
 	unsigned char *filedata;
-	long filesize;
+	int filesize;
 	section_t *sections;
 	int nsections;
 	int islib;				// set to true if the file is a "-l" option
